Rotation working, movement broken

+++ Acacia/Engine.h
@@ -58,7 +58,7 @@ private:
 	GLfloat cameraYaw;
 	GLfloat cameraPitch;
 	bool inputKeys[1024]; 
-	glm::vec2 mouseCurrentPosition;
+	glm::uvec2 mouseCurrentPosition;
 	GLfloat deltaTime = 0.0f;
 	GLfloat lastFrameTime = 0.0f;
 	GLfloat time;
